Clean up auth request state
